added .env.sample, refactored bot.py
