+++ app/.env.sample
@@ -1,0 +1,1 @@
+BOT_TOKEN=XXXXXXXXXXX
